Refactor semian_resource related functions
- Create sysv_semaphores to forward declare required prototypes
- Remove semian_ prefixes from file names
