+++ ext/semian/resource.h
@@ -3,12 +3,12 @@ For core semian resource functions exposed directly to ruby.
 
 Functions here are associated with rubyland operations.
 */
-
 #ifndef SEMIAN_RESOURCE_H
 #define SEMIAN_RESOURCE_H
 
-#include "semian.h" // FIXME remove this once temporary declarations are removed
+#include "semian.h" // FIXME remove this once configure_tickets has been refactored
 #include "types.h"
+#include "sysv_semaphores.h"
 
 // Ruby variables
 ID id_timeout;
