+++ ext/semian/semian.c
@@ -1,87 +1,11 @@
 #include "semian.h"
 
-// Time to wait for timed ops to complete
-#define INTERNAL_TIMEOUT 5 // seconds
-
-key_t
-generate_key(const char *name)
-{
-  char semset_size_key[20];
-  char *uniq_id_str;
-
-  // It is necessary for the cardinatily of the semaphore set to be part of the key
-  // or else sem_get will complain that we have requested an incorrect number of sems
-  // for the desired key, and have changed the number of semaphores for a given key
-  sprintf(semset_size_key, "_NUM_SEMS_%d", SI_NUM_SEMAPHORES);
-  uniq_id_str = malloc(strlen(name)+strlen(semset_size_key)+1);
-  strcpy(uniq_id_str, name);
-  strcat(uniq_id_str, semset_size_key);
-
-  union {
-    unsigned char str[SHA_DIGEST_LENGTH];
-    key_t key;
-  } digest;
-  SHA1((const unsigned char *) uniq_id_str, strlen(uniq_id_str), digest.str);
-  free(uniq_id_str);
-  /* TODO: compile-time assertion that sizeof(key_t) > SHA_DIGEST_LENGTH */
-  return digest.key;
-}
-
-void
-raise_semian_syscall_error(const char *syscall, int error_num)
-{
-  rb_raise(eSyscall, "%s failed, errno: %d (%s)", syscall, error_num, strerror(error_num));
-}
-
-void
-set_semaphore_permissions(int sem_id, long permissions)
-{
-  union semun sem_opts;
-  struct semid_ds stat_buf;
-
-  sem_opts.buf = &stat_buf;
-  semctl(sem_id, 0, IPC_STAT, sem_opts);
-  if ((stat_buf.sem_perm.mode & 0xfff) != permissions) {
-    stat_buf.sem_perm.mode &= ~0xfff;
-    stat_buf.sem_perm.mode |= permissions;
-    semctl(sem_id, 0, IPC_SET, sem_opts);
-  }
-}
-
-static const int kInternalTimeout = 5; /* seconds */
-
-static int
-get_max_tickets(int sem_id)
-{
-  int ret = semctl(sem_id, SI_SEM_CONFIGURED_TICKETS, GETVAL);
-  if (ret == -1) {
-    rb_raise(eInternal, "error getting max ticket count, errno: %d (%s)", errno, strerror(errno));
-  }
-  return ret;
-}
-
-int
-perform_semop(int sem_id, short index, short op, short flags, struct timespec *ts)
-{
-  struct sembuf buf = { 0 };
-
-  buf.sem_num = index;
-  buf.sem_op  = op;
-  buf.sem_flg = flags;
-
-  if (ts) {
-    return semtimedop(sem_id, &buf, 1, ts);
-  } else {
-    return semop(sem_id, &buf, 1);
-  }
-}
-
 static VALUE
 update_ticket_count(update_ticket_count_t *tc)
 {
   short delta;
   struct timespec ts = { 0 };
-  ts.tv_sec = kInternalTimeout;
+  ts.tv_sec = INTERNAL_TIMEOUT;
 
   if (get_max_tickets(tc->sem_id) != tc->tickets) {
     delta = tc->tickets - get_max_tickets(tc->sem_id);
@@ -91,7 +15,7 @@ update_ticket_count(update_ticket_count_t *tc)
     }
 
     if (semctl(tc->sem_id, SI_SEM_CONFIGURED_TICKETS, SETVAL, tc->tickets) == -1) {
-      rb_raise(eInternal, "error updating max ticket count, errno: %d (%s)", errno, strerror(errno));
+      rb_raise(eInternal, "error updating configured ticket count, errno: %d (%s)", errno, strerror(errno));
     }
   }
 
@@ -121,19 +45,19 @@ configure_tickets(int sem_id, int tickets, int should_initialize)
       while (get_max_tickets(sem_id) == 0) {
         usleep(10000); /* 10ms */
         gettimeofday(&cur_time, NULL);
-        if ((cur_time.tv_sec - start_time.tv_sec) > kInternalTimeout) {
+        if ((cur_time.tv_sec - start_time.tv_sec) > INTERNAL_TIMEOUT) {
           rb_raise(eInternal, "timeout waiting for semaphore initialization");
         }
       }
     }
 
     /*
-       If the current max ticket count is not the same as the requested ticket
+       If the current configured ticket count is not the same as the requested ticket
        count, we need to resize the count. We do this by adding the delta of
-       (tickets - current_max_tickets) to the semaphore value.
+       (tickets - current_configured_tickets) to the semaphore value.
     */
     if (get_max_tickets(sem_id) != tickets) {
-      ts.tv_sec = kInternalTimeout;
+      ts.tv_sec = INTERNAL_TIMEOUT;
 
       if (perform_semop(sem_id, SI_SEM_LOCK, -1, SEM_UNDO, &ts) == -1) {
         raise_semian_syscall_error("error acquiring internal semaphore lock, semtimedop()", errno);
@@ -152,42 +76,6 @@ configure_tickets(int sem_id, int tickets, int should_initialize)
       }
     }
   }
-}
-
-int
-create_semaphore(int key, long permissions, int *created)
-{
-  int semid = 0;
-  int flags = 0;
-
-  *created = 0;
-  flags = IPC_EXCL | IPC_CREAT | permissions;
-
-  semid = semget(key, SI_NUM_SEMAPHORES, flags);
-  if (semid >= 0) {
-    *created = 1;
-  } else if (semid == -1 && errno == EEXIST) {
-    flags &= ~IPC_EXCL;
-    semid = semget(key, SI_NUM_SEMAPHORES, flags);
-  }
-  return semid;
-}
-
-void *
-acquire_semaphore_without_gvl(void *p)
-{
-  semian_resource_t *res = (semian_resource_t *) p;
-  res->error = 0;
-  if (perform_semop(res->sem_id, SI_SEM_TICKETS, -1, SEM_UNDO, &res->timeout) == -1) {
-    res->error = errno;
-  }
-  return NULL;
-}
-
-int
-get_semaphore(int key)
-{
-  return semget(key, SI_NUM_SEMAPHORES, 0);
 }
 
 void Init_semian()
